requirements.txt
soundfile
sounddevice
numpy
torch
torchaudio
scikit-learn
librosa
argparse
pyyaml
tqdm

